Rename module which now contains a single widget
